Use separate files for package benchmarks
